backend/pyproject.toml
[tool.poetry]
name = "videosdb"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = ">=3.6.2,<3.7.0 || >3.7.0,<3.7.1 || >3.7.1,<4"
Django = "^3.1.4"
youtube-transcript-api = {git = "https://github.com/jdepoix/youtube-transcript-api.git"}
djangorestframework = "^3.12.2"
django-cors-headers = "^3.6.0"
django-uuslug = "^1.2.0"
httplib2 = "^0.18.1"
django-filter = "^2.4.0"
gunicorn = "^20.0.4"
django-extensions = "^3.1.0"
six = "^1.15.0"
django-test-without-migrations = "^0.6"
isodate = "^0.6.0"
ipfshttpclient = {git = "https://github.com/ipfs-shipyard/py-ipfs-http-client.git"}
executor = "^23.2"
youtube-dl = {git = "https://github.com/ytdl-org/youtube-dl.git"}
google-cloud-dns = "^0.32.1"
psycopg2 = "^2.9.1"
django-environ = "^0.7.0"

[tool.poetry.dev-dependencies]
ipdb = "^0.13.4"
pylint = "^2.6.0"
django-extensions = "^3.1.0"
autopep8 = "^1.5.4"
jedi = "<0.18"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
